requirements.txt
scrapy
faiss-cpu
numpy
apache-airflow


